frontend/small-fixes

+++ frontend/src/screens/Review.js
@@ -14,10 +14,12 @@ const Review = () => {
     const [rating, setRating] = useState(0)
     const [hover, setHover] = useState(0)
     const [fav, setFav] = useState(0)
-    const [date, setDate] = useState("")
+    const [date, setDate] = useState(new Date())
     const [show, setShow] = useState(false)
     const [comment, setComment] = useState("")
     const [cartoon, setCartoon] = useState({})
+    const [onWatchlist, setOnWatchlist] = useState(false)
+    const [isFav, setIsFav] = useState(false)
 
     axios.defaults.headers.common['Authorization'] = localStorage.getItem("token")
 
@@ -41,11 +43,20 @@ const Review = () => {
             comment: comment
         }
 
-        axios.delete(`${baseUrl}/users/watchlist/${params.id}`).then((res) => {
-
+        axios.get(`${baseUrl}/users/watchlist/${params.id}`).then((res) => {
+            if (res.data.found) {
+                setOnWatchlist(true)
+            }
         }).catch((err) => {
-            console.log(err)
         })
+
+        if (onWatchlist) {
+            axios.delete(`${baseUrl}/users/watchlist/${params.id}`).then((res) => {
+
+            }).catch((err) => {
+                console.log(err)
+            })
+        }
 
         axios.post(`${baseUrl}/reviews`, body).then((res) => {
 
@@ -64,7 +75,15 @@ const Review = () => {
             console.log(err)
         })
 
-        if (fav) {
+
+        axios.get(`${baseUrl}/users/watchlist/${params.id}`).then((res) => {
+            if (res.data.found) {
+                setIsFav(true)
+            }
+        }).catch((err) => {
+        })
+
+        if (fav && !isFav) {
             axios.patch(`${baseUrl}/users/favcartoons/${params.id}`).then((res) => {
 
             }).catch((err) => {
@@ -73,7 +92,8 @@ const Review = () => {
 
         }
 
-        if (!fav) {
+
+        if (!fav && isFav) {
             axios.delete(`${baseUrl}/users/favcartoons/${params.id}`).then((res) => {
 
             }).catch((err) => {
@@ -84,7 +104,7 @@ const Review = () => {
         body = {
             twatched: 1
         }
-        
+
         axios.patch(`${baseUrl}/users/update`, body).then((res) => {
 
         }).catch((err) => {
